requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
ibm-watson-machine-learning==1.0.335